This is the start of a major rework of sempr.
We are switching from the class hierarchy to an entity-component-system
architecture. As a first step I removed a lot of code that is obviously
no longer needed and added a Component-Class.

There is no functional code inside here currently. There is *a lot* to
do, and most of the remaining files need major reworks, too. I just left
them as a reminder on how they worked and what needs to be transferred.

Furthermore, I removed a lot of stuff from the CMakeLists to make the
project compile again, so we can incrementally add/rework parts.