fix loadFromJSON for multiple entries in archive
this->load(ar) directly tries to read the values, which fails if
multiple objects are present in the archive as it would need to descend
one level first. Calling ar(*this) solves this by letting cereal handle
the "next object incoming!"-part.

this->load(ar) works only with:

{
  myValue1 : "foo",
  myValue2 : "bar"
}

while ar(*this) supports this:

{
  component1 : {
    myValue1 : "foo",
    myValue2 : "bar"
  },
  component2 : {
    myValue1 : "baz",
    myValue2 : "foobar"
  }
}