changed EntityAccessor to inherit StringAccessor
The intention behind this is to allow an easy representation of an
entity as its id. That way we can write

[ (?e <type> <something>), EC<Component>(?e ?c) -> ... ]

and allow the GenericJoin to compare the two ?e accessors, one
pointing to the subject of a triple, the other to an Entity::Ptr.

The way it is implemented is a bit of a quick-hack. The whole accessor
system could use a major overhaul -- the current inheritance structure
is really weird, and inflexible.