Revert "gitlab-ci: install newer compiler (g++-8)"
This reverts commit c6b70c8f1586a8a2f80cfa460abdfd1539b9540e.

yes, g++8 fixes a compiler-bug, but it is not available on xenial.
Instead of worrying more about the compiler version, I implemented a
small workaround for the problem in the rete library.

Ultimatively we _should_ use a fixed compiler. But I did neither want to
abandon xenial yet, nor hassle with installing g++-8 in ubuntu 16.04 and
probably creating ABI incompatibilities when linking sempr there with
other libs/programs that use an older compiler.