encapsulated plugins in plugin builders
Some plugins might want to provide a "global variable", like a soprano-
oder other query-module. In the previous structure, those would have
been shared across multiple sempr::Core instances. To eliminate this,
the plugin stuff has been changed: The plugins themselves (now called
CapabilityBuilderInterface) only construct instances of
CapabilityInterface. For backward compatibility, the plugins are not
renamed. So e.g. you can still access the RDFPlugin by using
getPlugin<RDFPlugin>, and use it as usual, it is just not the plugin
anymore, but the capability constructed by the RDFPluginBuilder, which
is the actual plugin that is dynamically loaded.